+++ changelog.md
@@ -7,12 +7,13 @@
 - Add module setting: applyWeightMultiplierForProficientArmorHeavy
 - Add module setting: applyWeightMultiplierForProficientArmorNatural
 - Update constants for flags
-- Add flag veweight for see the updated weight on the item sheet with additionial check
+- Add flag veweight for see the updated weight on the item sheet with additional check
 - Add pre filter on wrapper method updateDocuments
 - Bug fix: https://github.com/p4535992/foundryvtt-variant-encumbrance-dnd5e/issues/24
 - Add module setting: applyWeightMultiplierForEquippedWeapon
 - Add module setting: applyWeightMultiplierForProficientWeapon
 - Add module setting: disableEquippedUnequippedProficientWeightManagement
+- Add some new hooks for customize the weight behaviour
 
 ### 0.13.12